# HRDepartment.Domain/HRDepartment.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HRDepartment.Api",
    "version": "1.0"
  },
  "paths": {
    "/BenefitType": {
      "get": {
        "tags": [
          "BenefitType"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BenefitTypeGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BenefitTypeGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BenefitTypeGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BenefitType"
        ],
        "summary": "Добавляет новый тип льготы.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenefitTypePostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BenefitTypePostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BenefitTypePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/BenefitType/{id}": {
      "get": {
        "tags": [
          "BenefitType"
        ],
        "summary": "Получает тип льготы по уникальному идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "BenefitType"
        ],
        "summary": "Обновляет информацию о существующем типе льготы.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenefitTypePostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BenefitTypePostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BenefitTypePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenefitTypeGetDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "BenefitType"
        ],
        "summary": "Удаляет тип льготы по уникальному идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Position": {
      "get": {
        "tags": [
          "Position"
        ],
        "summary": "Получает список всех должностей на предприятии.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Position"
        ],
        "summary": "Добавляет новую должность на предприятии.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PositionPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PositionGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionGetDto"
                }
              }
            }
          }
        }
      }
    },
    "/Position/{id}": {
      "get": {
        "tags": [
          "Position"
        ],
        "summary": "Получает должность по указанному идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PositionGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Position"
        ],
        "summary": "Обновляет информацию о должности по указанному идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PositionPostDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Position"
        ],
        "summary": "Удаляет должность по указанному идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BenefitTypeGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Уникальный идентификатор типа льготы",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название типа льготы",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для типа льготы"
      },
      "BenefitTypePostDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название типа льготы",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания нового типа льготы"
      },
...
      "PositionGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Уникальный идентификатор должности",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название должности",
            "nullable": true
          },
          "departmentId": {
            "type": "integer",
            "description": "Идентификатор отдела, к которому относится должность",
            "format": "int32"
          },
          "departmentName": {
            "type": "string",
            "description": "Название отдела, к которому относится должность",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для получения информации о должностях"
      },
      "PositionPostDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название должности",
            "nullable": true
          },
          "departmentId": {
            "type": "integer",
            "description": "Идентификатор отдела, к которому относится должность",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания новой должности"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
    }
  }
}